ml-api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas==2.1.1
scikit-learn==1.3.2
numpy==1.26.0
joblib==1.3.2
python-dotenv==1.0.0
httpx==0.25.0
